refactor: Reorganize package structure
